requirements.txt
torch==1.7.1
torchvision==0.8.2
timm==0.3.2
Pillow
blobfile
mypy
numpy
pytest
requests
einops
tensorboardX
deepspeed==0.4.0
scipy
omegaconf==2.0.0
pytorch-lightning==1.0.8